+++ BeetMall/src/main/java/com/beetmall/sshj/controller/FarmController.java
@@ -42,14 +42,7 @@ public class FarmController {
 	public String edit_farm() {
 		return "seller/edit_farm";
 	}
-	@RequestMapping("/notice")
-	public String notice() {
-		return "seller/notice";
-	}
-	@RequestMapping("/notice_view")
-	public String noticeView() {
-		return "seller/notice_view";
-	}	
+
 	@RequestMapping("/getListData") // seller_sales 페이지에서 차트와 엑셀 데이터 가져오는 dao
 	@ResponseBody
 	public List<FarmVO> getListData(FarmVO vo,HttpSession session) {
@@ -57,24 +50,5 @@ public class FarmController {
 		
 		return service.salesDataList(vo);
 	}
-	@RequestMapping("/ask_admin_list")
-	public String ask_admin_list() {
-		return "seller/ask_admin_list";
-	}
-	@RequestMapping("/ask_admin_view")
-	public String ask_admin_view() {
-		return "seller/ask_admin_view";
-	}
-	@RequestMapping("/ask_admin_write")
-	public String ask_admin_write() {
-		return "seller/ask_admin_write";
-	}
-	@RequestMapping("/ask_admin_edit")
-	public String ask_admin_edit() {
-		return "seller/ask_admin_edit";
-	}
-	@RequestMapping("/faq")
-	public String faq() {
-		return "seller/faq";
-	}
+	
 }
